requirements.txt
asgiref==3.2.3
django-ckeditor==5.9.0
django-jet==1.0.8
django-js-asset==1.2.2
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
